rename myAssignmentAction to assignmentActions
